ADR-002 stage 2.6 (PR3/3, final slice): split view/groups/pins/grid intents out of register_canvas
Relocates the last remaining closures out of register_canvas - Notes/
Frames/Containers/Groups, navigation pins, canvas view/behavior toggles
plus organize/font, and grid-control appearance - into four new
backend/api/intents_*.py modules, completing the split PR1 and PR2
started. Pure code motion: every closure is byte-identical to the
pre-PR3 baseline, with no comment changes needed this time (unlike PR1/
PR2, nothing here referenced code that moved to a different file).

Adds make_publish_grid to backend/api/_shared.py as publish_scene's
grid-control counterpart, and removes register_canvas's own now-dead
local publish_scene/publish_grid closures now that every consumer has
relocated.

register_canvas is now the thin orchestrator ADR-002 stage 2.6 set out
to produce: it constructs the shared SceneDocument, registers the four
topics with no natural feature-module home, and calls all 15
register_*_intents functions in turn. It registers zero intents
directly.
